Refactors the main server.js, and moves welcome msg to separate file
